Fixed assign bug resetting values and created InteractsWithLists trait
